+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### **Added**
 
+- added support for whitelisting custom CIDRs from SSM Parameters and AWS Codebuild IPs for being able to run seedfarmer commands
+
 ### **Changed**
 
 ### **Removed**